backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
tensorflow==2.20.0
numpy==1.26.0
pillow==10.0.0
python-multipart==0.0.6
pytest==7.4.0
httpx==0.24.0
